requirements.txt
coverage==4.5.3
Django==2.2
djangorestframework==3.9.3
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
